src/aisprint/drift_detector/requirements.txt
numpy==1.23
alibi-detect
influxdb_client
boto3
PyYAML